pyproject.toml
[project]
name = "NTSAgent"
version = "0.0.19"
description = "사업자등록번호로 사업자 등록 상태를 체크하는 에이전트"
authors = [{ name = "Charles Kim", email = "hyungsoo.kim@uipath.com" }]
dependencies = [
    "langgraph>=0.2.70",
    "langchain-core>=0.3.34",
    "langchain-openai>=0.3.9",
    "langgraph-checkpoint-sqlite>=2.0.3",
    "python-dotenv>=1.0.1",
    "uipath-langchain==0.0.100",
    "pydantic>=2.10.6",
    "aiohttp>=3.11.12",
    "typing-extensions>=4.12.2",
    "ipython>=8.32.0",
    "langchain-anthropic>=0.3.13",
]
requires-python = ">=3.10"
